chore: Add DelegationSet to Status for HostedZone (#64)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/2364

Description of changes:

Add the DelegationSet to the Status of HostedZone resources so the nameservers can be retrieved to then be referenced in a RecordSet for a different HostedZone.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

I'm not sure if there's a more proper way to get the sdkCreate and sdkFind calls to be generated with the relevant logic without the hook but I couldn't find a way to get that to happen. As such I've gone down the hook path.